requirements.txt
# IndiPaperTrade Requirements

# Angel One SmartAPI
smartapi-python

# TOTP generation for SmartAPI authentication
pyotp

# Data manipulation and analysis
pandas
numpy

# WebSocket support
websocket-client

# Dashboard / UI
streamlit
PyQt6

# Database ORM and utilities
sqlalchemy

# Environment variables
python-dotenv

# Timezone support
pytz

# HTTP requests
requests

# JSON handling (optional, standard library but good to be explicit)
# json is built-in

# Logging (optional, standard library)
# logging is built-in

# Threading (optional, standard library)
# threading is built-in

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# pylint>=2.0.0
# mypy>=0.900
